commit
make seeder

+++ database/migrations/2023_12_5_144051_create_jadwals_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('jadwal', function (Blueprint $table) {
-            $table->unsignedBigInteger('id_jadwal');
+            $table->integer('id_jadwal');
             $table->primary('id_jadwal');
             $table->unsignedBigInteger('id_sesi');
             $table->unsignedInteger('kode_pelatih');
